Update ExternalSearchAnnotationSidebar.java
Each time the search result is reloaded, a new ResultRowView type object is generated. We need to add statements in the definition of the ResultRowView class to determine whether the search result of the left sidebar is clicked. searchState.getSelectedResult() == null. Before without clicking  the search result of the sidebar on the left side, add the title content to the link, otherwise you need to use if(documentTitle.equals(rowView.title)) to determine which of the 8 search results is the same as the title value of the current click result. Returning true search results's title value is highlighted and added to the link.Returning false directly add to the link without highlights. The following statement is used for the highlight title value: documentTitle="<mark>"+documentTitle+"</mark>";
rowView.link.add(newLabel("title",documentTitle).setEscapeModelStrings(false));